Improve contracts
