+++ packages/contracts/source/contracts/p2p/header-service.ts
@@ -1,0 +1,7 @@
+import { IHeader, IHeaderData } from "./header";
+import { Peer } from "./peer";
+
+export interface IHeaderService {
+	getHeader(): IHeader;
+	handle(peer: Peer, header: IHeaderData): Promise<void>;
+}
