+++ packages/contracts/source/contracts/p2p/header.ts
@@ -1,13 +1,16 @@
-import { IHeaderData, Peer } from "./peer";
+export type IHeaderData = {
+	version: string;
+	height: number;
+	round: number;
+	step: number;
+	proposedBlockId: string | null;
+	validatorsSignedPrevote: boolean[];
+	validatorsSignedPrecommit: boolean[];
+};
 
 export interface IHeader {
 	toData(): IHeaderData;
 	// hasHigherHeight(headerData: IHeaderData): boolean;
 	// hasMissingProposal(headerData: IHeaderData): boolean;
 	// hasMissingMessages(headerData: IHeaderData): boolean;
-}
-
-export interface IHeaderService {
-	getHeader(): IHeader;
-	handle(peer: Peer, header: IHeaderData): Promise<void>;
 }
